app/requirements.txt
absl-py==0.8.1
bleach==1.5.0
certifi==2019.9.11
Click==7.0
cycler==0.10.0
Flask==1.1.1
html5lib==0.9999999
itsdangerous==1.1.0
Jinja2==2.10.3
kiwisolver==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.1
numpy==1.17.4
Pillow==6.2.1
protobuf==3.10.0
pyparsing==2.4.5
python-dateutil==2.8.1
six==1.13.0
terminaltables==3.1.0
torch==1.1.0
torchvision==0.3.0
tqdm==4.38.0
Werkzeug==0.16.0
wincertstore==0.2
